+++ lib/src/platform_icons.dart
@@ -215,10 +215,12 @@ class PlatformIcons {
   /// Icons: Icons.more_horiz : CupertinoIcons.ellipsis
   IconData get ellipsis =>
       isMaterial(context) ? Icons.more_horiz : CupertinoIcons.ellipsis;
-  
+
   /// Icons: Icons.error : CupertinoIcons.(custom)
   IconData get error => isMaterial(context)
-      ? Icons.error : IconData(0xf35b, fontFamily: 'CupertinoIcons', fontPackage: 'cupertino_icons');
+      ? Icons.error
+      : IconData(0xf35b,
+          fontFamily: 'CupertinoIcons', fontPackage: 'cupertino_icons');
 
   // IconData get eye => isMaterial(context) ? Icons. : CupertinoIcons.eye;
 
@@ -292,14 +294,18 @@ class PlatformIcons {
   /// Icons: Icons.favorite : CupertinoIcons.heart_solid
   IconData get heartSolid =>
       isMaterial(context) ? Icons.favorite : CupertinoIcons.heart_solid;
-  
+
   /// Icons: Icons.help : CupertinoIcons.(custom)
   IconData get help => isMaterial(context)
-      ? Icons.help : IconData(0xf446, fontFamily: 'CupertinoIcons', fontPackage: 'cupertino_icons');
-  
-  /// Icons: Icons.help_outline : CupertinoIcons.(custom)
-  IconData get help_outline => isMaterial(context)
-      ? Icons.help_outline : IconData(0xf445, fontFamily: 'CupertinoIcons', fontPackage: 'cupertino_icons');
+      ? Icons.help
+      : IconData(0xf446,
+          fontFamily: 'CupertinoIcons', fontPackage: 'cupertino_icons');
+
+  /// Icons: Icons.helOutline : CupertinoIcons.(custom)
+  IconData get helpOutline => isMaterial(context)
+      ? Icons.help_outline
+      : IconData(0xf445,
+          fontFamily: 'CupertinoIcons', fontPackage: 'cupertino_icons');
 
   /// Icons: Icons.home : CupertinoIcons.home
   IconData get home => isMaterial(context) ? Icons.home : CupertinoIcons.home;
@@ -526,7 +532,7 @@ class PlatformIcons {
 
   // IconData get shuffleThick =>
   //     isMaterial(context) ? Icons.share : CupertinoIcons.shuffle_thick;
-  
+
   /// Icons: Icons.star : CupertinoIcons.(custom)
   IconData get star => isMaterial(context)
       ? Icons.star